Made modifications to mutexes.c & timedmutexes.c for display purposes.  New display will print timestamp followd by thread id within [].
Additional changes include adding "valid" value to attitude
struct.  This is used by the READER thread when it obtains
the lock before the UPDATE thread has been able to write
valid data.  In the mutexes process, this means the READER
thread might run two times if it wins the race for the lock.
Likewise int the timedmutexes process, it would only come
into play if the READER thread was to win the race to the
lock after creation.

I also moved many of the common items used by each to a new
header file to help cut down on make changes to both files.

The total number of print statments seems a little overkill,
however, it does clearly represent what each process is
doing. Feel free to cut some out if you want.